% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_docs.R
\name{docs_tab}
\alias{docs_tab}
\title{Documentation tab UI}
\usage{
docs_tab()
}
\value{
A \code{shiny.tabPanel}.
}
\description{
Builds the Documentation tab of the ADMETShiny application.
}
\keyword{internal}
