% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_tutorial.R
\name{tutorial_tab}
\alias{tutorial_tab}
\title{Tutorial tab UI}
\usage{
tutorial_tab()
}
\value{
A \code{shiny.tabPanel}.
}
\description{
Builds the Tutorial tab of the ADMETShiny application.
}
\keyword{internal}
