% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admetshiny-package.R
\docType{package}
\name{admetshiny-package}
\alias{admetshiny}
\alias{admetshiny-package}
\title{admetshiny: Interactive ADMET and Drug-Likeness Analysis}
\description{
The \pkg{admetshiny} package provides an interactive Shiny application and a
collection of R functions for the management, calculation, filtering,
visualization and exploratory analysis of molecular descriptors and ADMET
properties of small molecules.
computes descriptors locally via the Chemistry
Development Kit (CDK) through \pkg{rcdk}, and offers drug-likeness filters
(Lipinski, Veber, Ghose, Egan, Muegge), the BOILED-Egg model, PCA, t-SNE,
radar plots and Tanimoto/AGNES structural similarity clustering.
}
\details{
To launch the interactive application simply run:

\preformatted{
admetshiny::run_app()
}

The exported functions (filters, plots, data fixers and CDK helpers) can
also be used programmatically in plain R scripts.
By Xavier Clemente Garcia Cevallos :3 "Hecho con amor para la ciencia"
}
\section{Drug-likeness filters}{


\itemize{
\item \code{\link{lipinskiFilter}} - Lipinski Rule-of-Five
\item \code{\link{veberFilter}} - Veber oral bioavailability
\item \code{\link{ghoseFilter}} - Ghose qualifying range
\item \code{\link{eganFilter}} - Egan PSA-LogP rule
\item \code{\link{mueggeFilter}} - Muegge pharmacophore filter
\item \code{\link{applyFilters}} - Apply multiple filters in sequence
\item \code{\link{computeViolationColumns}} - Compute violation columns
}
}

\section{Data normalization}{


\itemize{
\item \code{\link{mapADMETColumns}} - Map any user dataset to standard schema
\item \code{\link{mapCDKDescriptors}} - Map CDK descriptors to standard schema
}
}

\section{CDK & webchem}{


\itemize{
\item \code{\link{getSmilesFromIdentifiers}} - Retrieve SMILES from PubChem
\item \code{\link{calcCDKDescriptors}} - Calculate descriptors with CDK
\item \code{\link{computeADMETProperties}} - Compute BOILED-Egg ADMET properties
}
}

\section{Visualizations}{


\itemize{
\item \code{\link{plotBoiledEgg}} - BOILED-Egg model
\item \code{\link{plotMW}}, \code{\link{plotTPSA}}, \code{\link{plotLogP}} - Distributions
\item \code{\link{plotPCA}} - PCA chemical space
\item \code{\link{plotTSNE}} - t-SNE chemical space
\item \code{\link{plotUMAP}} - UMAP chemical space
\item \code{\link{plotParallel}} - Parallel coordinates
\item \code{\link{plotViolin}} - Violin plot
\item \code{\link{plotRadar}} - Radar chart
\item \code{\link{plotTanimoto}} - Tanimoto/AGNES clustering
\item \code{\link{plotCorrHeatmap}} - Correlation heatmap
\item \code{\link{plotClusterHeatmap}} - Cluster heatmap with dendrogram
\item \code{\link{plotHistogramCustom}} - Customizable histogram of any numeric column
\item \code{\link{palette_selector_ui}}, \code{\link{apply_palette}} - Colour customization
}
}

\section{Application}{


\itemize{
\item \code{\link{run_app}} - Launch the Shiny application
\item \code{\link{generateReport}} - Generate a report from the R console
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/xavierclementegarcia/admetshiny}
  \item Report bugs at \url{https://github.com/xavierclementegarcia/admetshiny-package/issues}
}

}
\author{
\strong{Maintainer}: Xavier Clemente Garcia Cevallos \email{xgarcia@unicauca.edu.co} (\href{https://orcid.org/0000-0002-3284-2363}{ORCID})

Authors:
\itemize{
  \item Xavier Clemente Garcia Cevallos \email{xgarcia@unicauca.edu.co} (\href{https://orcid.org/0000-0002-3284-2363}{ORCID})
}

}
